refactor(create-idea): fix just leader delete of change idea, refactor(api:accept invite): change status of user when accept invite

+++ server/src/services/idea.service.ts
@@ -97,9 +97,17 @@ export class IdeaService {
       }
     })
   }
-  async deleteIdea(projectId: string) {
+  async deleteIdea(projectId: string,userId: string) {
     return runTransaction(async (session) => {
+      const project = await this.projectModel.findOne({ _id: projectId }).session(session).exec()
+      if (project?.leader !== userId) {
+        throw new HttpException('You are not the leader of this idea', 400)   
+      }
       await this.projectModel.deleteOne({ _id: projectId }).session(session).exec()
+      await this.userModel.updateOne(
+        { _id: userId},
+        { $set: { status: USER_STATUS.UN_GROUPED } },        
+      ).session(session).exec()
     })
   }
 }
